refactor: Improve name of main implementation file.
